+++ js/data.js
@@ -1,0 +1,40 @@
+const NAMES = [
+  'Кристина',
+  'Игорь',
+  'Зина',
+  'Петруша',
+  'Макс',
+  'Кирилл',
+  'Алекс',
+  'Коля',
+  'Ксения',
+  'Алина',
+];
+const DESCRIPTIONS = [
+  'Солнечный берег с белым песком и пальмами.',
+  'Ночной город с огнями и движением автомобилей.',
+  'Улыбающиеся дети, играющие в парке.',
+  'Закат над горами с яркими оранжевыми и красными оттенками.',
+  'Стейк на гриле с овощами на стороне.',
+  'Старый замок на вершине холма, окруженный туманом.',
+  'Лесная тропинка среди осенних деревьев с разноцветной листвой.',
+  'Пара, прогуливающаяся по мосту, держа друг друга за руки.',
+  'Домашний офис с приятной атмосферой и растениями.',
+  'Уютный кафе с чашкой кофе и книгой на столе.',
+  'Дети, строящие снеговика на зимнем дворе.',
+  'Спокойное озеро с отражением облаков на воде.',
+  'Уличный рынок с яркими фруктами и овощами.',
+  'Групповой снимок друзей на пляже.',
+  'Кошка, спящая на окне в солнечный день.',
+];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо.Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.Как можно было поймать такой неудачный момент ? !',
+];
+const PHOTOS_COUNT = 25;
+
+export { NAMES, DESCRIPTIONS, MESSAGES, PHOTOS_COUNT };
